Put .htaccess in logs folder when inside wp-content
When the log root folder is inside WordPress content folder, the logs
are going to be publicly accessible, and that is in best case a privacy
leakage issue, in worst case a security threat.
We try to write an .htaccess file to prevent access to them.
This guarantees nothing, because .htaccess can be ignored depending web
server in use and its configuration, but at least we tried.