Add Query Monitor handler.
As Query Monitor is so ubiquitous througout WordPress development and it has it's
own [neat logging capability(https://querymonitor.com/docs/logging-variables/) it's useful to be able to configure Wonolog to write to the query monitor log.

This enables us to use the same logging code across environments and log
to query monitor on the environments we wish. For example, we log to
disk for local dev, QM on QA and Staging and New Relic in production.

Handlers should generally be added in userland, however as mentioned
with Query Monitor so ubiquitous it may be a nice offering to offer as
part of Wonolog.