restore_exception_handler does not work from an handler
Even if it never worked, in PHP < 8.3.0 throwing from an handler
results in the default handler to be called.

From PHP >= 8.3.0 < 8.3.5 throwing from an handler
results in the default previous hander to be called, if any.

From PHP >= 8.3.5 throwing from an handler would
result in the same handler to be called again and again, causing an
infinite loop.

To have a uniform behavior, we call manually the previous handler, if it
existed, then we call `set_exception_handler(null)` to force the default
handler, and finally we throw the exception.

This should ensure a same flow in all supported versions:

1. Log the exception
2. Execute the previous handler, if any
3. Re-throw the exception